Fix Small Error In DynamicQuantizeLinear Desc
The DynamicQuantizeLinear description was modified to reflect the computation that is actually performed. Namely, the zero-point should be the difference between the minimum value of the data and the minimum of the quantization range scaled using 'y_scale'; and not scaled using the width of the quantization range. This can also be seen from the example where it is done correctly.